Add email checkeer and set email address only if it is valid
